pyproject.toml
[project]
name = "linkedin-job-filter-backend"
version = "0.1.0"
description = "Backend for the LinkedIn Job Filter Chrome extension."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "supabase>=2.9",
    "httpx>=0.27",
    "anthropic>=0.39",
    "openai>=1.54",
    "pyjwt[crypto]>=2.9",
    "python-json-logger>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.32",
    "ruff>=0.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
